+++ src/vstboard/connectables/audiodevice_pa.cpp
@@ -625,7 +625,7 @@ float AudioDevice::GetCpuUsage()
 }
 
 #ifdef CIRCULAR_BUFFER
-bool AudioDevice::DeviceToRingBuffers( const void *inputBuffer, unsigned long framesPerBuffer)
+bool AudioDevice::DeviceToRingBuffers( const void *inputBuffer, qint32 framesPerBuffer)
 {
     //if no input, render when the host buffer is full
     static ulong fakeFilledSize=0;
@@ -635,7 +635,7 @@ bool AudioDevice::DeviceToRingBuffers( const void *inputBuffer, unsigned long fr
         if(isClosing) return false;
     }
 
-    unsigned long hostBuffSize = myHost->GetBufferSize();
+    qint32 hostBuffSize = myHost->GetBufferSize();
     if(framesPerBuffer > hostBuffSize) {
 #ifdef DEBUG_DEVICES
         LOG(QString("ajusting host buffer %1->%2")
@@ -724,7 +724,7 @@ void AudioDevice::PinsToRingBuffers()
 
 }
 
-bool AudioDevice::RingBuffersToDevice( void *outputBuffer, unsigned long framesPerBuffer)
+bool AudioDevice::RingBuffersToDevice( void *outputBuffer, qint32 framesPerBuffer)
 {
     {
         QMutexLocker lo(&mutexOpenClose);
@@ -801,9 +801,9 @@ bool AudioDevice::RingBuffersToDevice( void *outputBuffer, unsigned long framesP
     return true;
 }
 #else
-bool AudioDevice::DeviceToPinBuffers( const void *inputBuffer, unsigned long framesPerBuffer )
+bool AudioDevice::DeviceToPinBuffers( const void *inputBuffer, qint32 framesPerBuffer )
 {
-//    unsigned long hostBuffSize = myHost->GetBufferSize();
+//    qint32 hostBuffSize = myHost->GetBufferSize();
 //    if(framesPerBuffer > hostBuffSize) {
 //       myHost->SetBufferSize(framesPerBuffer);
 //       hostBuffSize = framesPerBuffer;
@@ -835,9 +835,9 @@ bool AudioDevice::DeviceToPinBuffers( const void *inputBuffer, unsigned long fra
     return true;
 }
 
-bool AudioDevice::PinBuffersToDevice( void *outputBuffer, unsigned long framesPerBuffer )
+bool AudioDevice::PinBuffersToDevice( void *outputBuffer, qint32 framesPerBuffer )
 {
-//    unsigned long hostBuffSize = myHost->GetBufferSize();
+//    qint32 hostBuffSize = myHost->GetBufferSize();
 //    if(framesPerBuffer > hostBuffSize) {
 //       myHost->SetBufferSize(framesPerBuffer);
 //       hostBuffSize = framesPerBuffer;
@@ -852,7 +852,7 @@ bool AudioDevice::PinBuffersToDevice( void *outputBuffer, unsigned long framesPe
 
         for (int i = 0; i < devInfo.maxOutputChannels; i++) {
         // for (int i = 0; i < objInfo.outputs; i++) {
-            for (unsigned long j = 0; j < framesPerBuffer; j++) {
+            for (qint32 j = 0; j < framesPerBuffer; j++) {
 				((float **)outputBuffer)[i][j] = .0f;
 			}
 		}
